optical_flow/requirements.txt
pyyaml
openai
requests
dotenv
opencv-python
opencv-contrib-python
modelscope
pillow
openai-whisper